Scale down instances, increase cpu limit
In an attempt to speed things up for now, the cpu limit is once again increased but the number of instances is now down to 1.  I want to do a few other things in the cluster to save money before I add another cluster node which seems to be what needs to happen.